LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/trigger-scope-in-scope-source.tentative-expected.txt

In-scope Target
SOURCE
Out-of-scope Target

FAIL In-scope target finds trigger of source; out-of-scope target does not. assert_equals: animation on in_scope_target has playState paused. expected "paused" but got "running"
